test: apply tests
