+++ packages/sdk/src/index.ts
@@ -11,5 +11,5 @@ import { handleError } from "./error";
 const initVSFSDK = initSDK;
 
 export { initVSFSDK, initSDK, handleError, buildModule, eventManager };
-export * from "./modules/middlewareModule";
+export * from "./modules/connect";
 export * from "./types";
