+++ packages/sdk/src/modules/connect/consts.ts
@@ -1,0 +1,4 @@
+/**
+ * Symbol to mark a method config object.
+ */
+export const isRequestConfig = Symbol("methodConfig");
