+++ packages/sdk/src/modules/connect/index.ts
@@ -1,0 +1,3 @@
+export * from "./types";
+export * from "./utils";
+export { connect } from "./module";
