+++ packages/sdk/src/modules/connect/utils/index.ts
@@ -1,0 +1,2 @@
+export { getHttpClient } from "./getHttpClient";
+export { prepareConfig } from "./prepareConfig";
